Move widgets to the react-jsonschema-form-bootstrap, init theme
