+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/CheckboxesWidget.js
