+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/ColorWidget.js
