+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/EmailWidget.js
