+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/HiddenWidget.js
