+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/SelectWidget.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { asNumber } from '../../utils';
+import { asNumber } from 'react-jsonschema-form/src/utils';
 
 /**
  * This is a silly limitation in the DOM where option change event values are
