+++ packages/react-jsonschema-form-bootstrap/src/components/widgets/TextWidget.js
