+++ packages/react-jsonschema-form-bootstrap/src/index.js
@@ -1,0 +1,3 @@
+import widgets from './components/widgets';
+
+export default { widgets };
